+++ src/PlaywrightRunner.ts
@@ -9,72 +9,42 @@ import type {
   TestRunnerOptions,
 } from 'jest-runner'
 import type { Config as JestConfig } from '@jest/types'
-import type { BrowserType } from './types'
-import {
-  checkBrowserEnv,
-  checkDeviceEnv,
-  getDisplayName,
-  readConfig,
-  getPlaywrightInstance,
-  readPackage,
-} from './utils'
+import { getDisplayName } from './utils'
 import { DEFAULT_TEST_PLAYWRIGHT_TIMEOUT } from './constants'
 
-const getBrowserTest = (
-  test: Test,
-  browser: BrowserType,
-  device: string | null,
-): Test => {
-  const { displayName } = test.context.config
-  const playwrightDisplayName = getDisplayName(browser, device)
-  return {
-    ...test,
-    context: {
-      ...test.context,
-      config: {
-        ...test.context.config,
-        // @ts-ignore
-        browserName: browser,
-        device,
-        displayName: {
-          name: displayName
-            ? `${playwrightDisplayName} ${
-                typeof displayName === 'string' ? displayName : displayName.name
-              }`
-            : playwrightDisplayName,
-          color: 'yellow',
+const getTests = (tests: Test[]): Test[] => {
+  //@ts-ignore
+  return tests.map((test) => {
+    const { displayName } = test.context.config
+    //@ts-ignore
+    const playwrightDisplayName = getDisplayName(test.browser, test.device)
+    const pwTest = {
+      ...test,
+      context: {
+        ...test.context,
+        config: {
+          ...test.context.config,
+          //@ts-ignore
+          browserName: test.browser,
+          displayName: {
+            name: displayName
+              ? `${playwrightDisplayName} ${
+                  typeof displayName === 'string'
+                    ? displayName
+                    : displayName.name
+                }`
+              : playwrightDisplayName,
+            color: 'yellow',
+          },
         },
       },
-    },
-  }
-}
-
-const getTests = async (tests: Test[]): Promise<Test[]> => {
-  const playwrightPackage = await readPackage()
-  const pwTests: Test[] = []
-  await Promise.all(
-    tests.map(async (test) => {
-      const { rootDir } = test.context.config
-      const { browsers, devices } = await readConfig(rootDir)
-      browsers.forEach((browser) => {
-        checkBrowserEnv(browser)
-        const { devices: availableDevices } = getPlaywrightInstance(
-          playwrightPackage,
-          browser,
-        )
-        if (devices && devices.length) {
-          devices.forEach((device) => {
-            const availableDeviceNames = Object.keys(availableDevices)
-            checkDeviceEnv(device, availableDeviceNames)
-            pwTests.push(getBrowserTest(test, browser, device))
-          })
-        } else {
-          pwTests.push(getBrowserTest(test, browser, null))
-        }
-      })
-    }),
-  )
-  return pwTests
+    }
+    //@ts-ignore
+    delete pwTest.browserName
+    //@ts-ignore
+    delete pwTest.device
+    return pwTest
+  })
 }
 
 class PlaywrightRunner extends JestRunner {
